requirements.txt
docopt
pyyaml
jsonmerge